Fix a double-send crash when no additional deps
E.g. /deps/RSEIS,RPMG,Rwave
The first depends on the other two, but they
don't depend on anything else, so there is
nothing to add, and we return early. But the
async loop already started by this, and when
it finishes it also sends the result.

We have a flag now that effectively cancels
the send of the async loop.